Implement createOrder in Angular to create order in database

+++ ECommerce/Controllers/OrderController.cs
@@ -55,23 +55,33 @@ namespace ECommerce.Controllers
             {
 
                 var address = await _addressRepository.GetByIdAsync(orderDto.AddressId);
-
+                if (address == null)
+                {
+                    return BadRequest(new { Message = "Address Not Found" });
+                }
                 var user = await _userManager.GetUserAsync(User);
-              
-                var order = await _orderService.CreateOrderAsync(user.Email, orderDto.CartId, orderDto.ShippingCostId,orderDto.AddressId);
+                if(user == null)
+                {
+                    return BadRequest(new { Message = "User Not Found" });
+                }
+                var order = await _orderService.CreateOrderAsync(user.Email, orderDto.CartId, orderDto.ShippingCostId,orderDto.AddressId,orderDto.PaymentId);
                 if (order == null)
                 {
                     return BadRequest(new { Message = "Failure in Order Process" });
                 }
                 var deletedcart = await _cartRepository.DeleteCartAsync(orderDto.CartId);
-                return Ok(new { Message= "Order Created Successfully" });
+                if (!deletedcart)
+                {
+                    return BadRequest(new { Message = "Failure in Cart Deletion" });
+                }
+                return Ok(new { Message= "Order Created Successfully",orderId=order.Id });
             }
             catch (Exception ex)
             {
 
                 _logger.LogError(ex, "Error occurred while creating the order.");
 
-                return BadRequest(new { Message = "failed!" });
+                return BadRequest(new { Message = "Failed to create order.", Error = ex.Message });
             }
         }
